DNS same hostname (#314)
* mt.sh: remove not-always-valid message
* haraka: only create dcc.ini once
* dns: no duplicate TXT when HOSTNAME=MAIL_DOMAIN

+++ provision-haraka.sh
@@ -600,6 +600,10 @@ EO_WL
 
 configure_haraka_dcc()
 {
+	if [ -f "$HARAKA_CONF/dcc.ini" ]; then
+		return
+	fi
+
 	tell_status "configuring DCC"
 	tee -a "$HARAKA_CONF/dcc.ini" <<EO_DCC
 [dccifd]
